chore: Delete unused files and move files #14
