--- meteor-viewmodel/src/androidMain/kotlin/io/spherelabs/meteorviewmodel/Platform.kt
@@ -1,7 +1,0 @@
-package io.spherelabs.meteorviewmodel
-
-class AndroidPlatform : Platform {
-    override val name: String = "Android ${android.os.Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()
