--- meteor-viewmodel/src/commonMain/kotlin/io/spherelabs/meteorviewmodel/Cancellable.kt
@@ -1,5 +1,0 @@
-package io.spherelabs.meteorviewmodel
-
-interface Cancellable {
-    fun cancel()
-}
