+++ meteor-viewmodel/src/commonMain/kotlin/io/spherelabs/meteorviewmodel/Closeable.kt
@@ -1,4 +1,4 @@
-package io.spherelabs.meteorviewmodel.core
+package io.spherelabs.meteorviewmodel
 
 /**
  * A [Closeable] manages resources that need to be closed or released when the ViewModel is no longer needed.
