--- meteor-viewmodel/src/commonMain/kotlin/io/spherelabs/meteorviewmodel/Platform.kt
@@ -1,7 +1,0 @@
-package io.spherelabs.meteorviewmodel
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
