+++ meteor-viewmodel/src/iosMain/kotlin/io/spherelabs/meteorviewmodel/Closeable.kt
@@ -1,0 +1,5 @@
+package io.spherelabs.meteorviewmodel
+
+public actual fun interface Closeable {
+    public actual fun close()
+}
