--- meteor-viewmodel/src/iosMain/kotlin/io/spherelabs/meteorviewmodel/Platform.kt
@@ -1,9 +1,0 @@
-package io.spherelabs.meteorviewmodel
-
-import platform.UIKit.UIDevice
-
-class IOSPlatform: Platform {
-    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
-
-actual fun getPlatform(): Platform = IOSPlatform()
